+++ horde/classes/base/waiting_prompt.py
@@ -19,7 +19,7 @@ from horde.classes.kobold.processing_generation import TextProcessingGeneration
 from horde.classes.stable.processing_generation import ImageProcessingGeneration
 from horde.flask import SQLITE_MODE, db
 from horde.logger import logger
-from horde.utils import get_db_uuid, get_expiry_date
+from horde.utils import get_db_uuid, get_expiry_date, get_extra_slow_expiry_date
 
 procgen_classes = {
     "template": ProcessingGeneration,
@@ -105,6 +105,7 @@ class WaitingPrompt(db.Model):
     things = db.Column(db.BigInteger, default=0, nullable=False)
     total_usage = db.Column(db.Float, default=0, nullable=False)
     extra_priority = db.Column(db.Integer, default=0, nullable=False, index=True)
+    # TODO: Delete. Obsoleted.
     job_ttl = db.Column(db.Integer, default=150, nullable=False)
     disable_batching = db.Column(db.Boolean, default=False, nullable=False)
     webhook = db.Column(db.String(1024))
@@ -204,7 +205,6 @@ class WaitingPrompt(db.Model):
         self.things = 0
         self.total_usage = round(self.things * self.n, 2)
         self.prepare_job_payload()
-        self.set_job_ttl()
         db.session.commit()
 
     def prepare_job_payload(self):
@@ -241,7 +241,7 @@ class WaitingPrompt(db.Model):
         self.n -= safe_amount
         payload = self.get_job_payload(current_n)
         # This does a commit as well
-        self.refresh()
+        self.refresh(worker)
         procgen_class = procgen_classes[self.wp_type]
         gens_list = []
         model = None
@@ -457,8 +457,13 @@ class WaitingPrompt(db.Model):
         except Exception as err:
             logger.warning(f"Error when aborting WP. Skipping: {err}")
 
-    def refresh(self):
-        self.expiry = get_expiry_date()
+    def refresh(self, worker=None):
+        if self.n > 0 and worker is not None and worker.extra_slow_worker is True:
+            self.expiry = get_extra_slow_expiry_date()
+        else:
+            new_expiry = get_expiry_date()
+            if self.expiry < new_expiry:
+                self.expiry = new_expiry
         db.session.commit()
 
     def is_stale(self):
@@ -468,13 +473,6 @@ class WaitingPrompt(db.Model):
 
     def get_priority(self):
         return self.extra_priority
-
-    def set_job_ttl(self):
-        """Returns how many seconds each job request should stay waiting before considering it stale and cancelling it
-        This function should be overriden by the invididual hordes depending on how the calculating ttl
-        """
-        self.job_ttl = 150
-        db.session.commit()
 
     def refresh_worker_cache(self):
         worker_ids = [worker.worker_id for worker in self.workers]